shared_python/unified-config-loader/pyproject.toml
[project]
name = "unified-config-loader"
version = "1.0.0"
description = "Pluggable configuration loader with env/vault/aws backends"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Lenie AI Team" }
]

dependencies = [
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
vault = ["hvac>=2.0.0"]
aws = ["boto3>=1.26.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["unified_config_loader"]

[dependency-groups]
dev = [
    "pytest",
    "ruff",
]

[tool.ruff]
line-length = 120
exclude = [".git", "__pycache__", ".venv"]

[tool.ruff.lint]
select = ["E", "F", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
